[DEBT-043-D6] Split agent/research_agent.py into focused modules
- Created agent/research_agent/ package with 5 files:
  - __init__.py (9 LoC): Re-exports ResearchAgent
  - _agent.py (236 LoC): Main ResearchAgent class with public API
  - _executor.py (166 LoC): Step execution logic
  - _plan_builder.py (152 LoC): Research plan building
  - _recovery.py (233 LoC): Deep research crash recovery
- All files under 400 LoC ceiling
- Updated test fixture to propagate isolated state to executor/recovery
- Updated test monkeypatches for new execution path
- Quality gates: pre-commit, mypy, pytest (1052 passed)
