+++ src/kalshi_research/agent/research_agent/__init__.py
@@ -1,0 +1,9 @@
+"""Research agent package for cost-bounded, deterministic research automation.
+
+This package provides the ResearchAgent class that orchestrates Exa API calls
+in a predictable, budget-aware manner.
+"""
+
+from kalshi_research.agent.research_agent._agent import ResearchAgent
+
+__all__ = ["ResearchAgent"]
